+++ backend/src/main/java/com/woowacourse/thankoo/coupon/application/CouponQueryService.java
@@ -6,8 +6,8 @@ import com.woowacourse.thankoo.coupon.domain.CouponStatus;
 import com.woowacourse.thankoo.coupon.domain.CouponStatusGroup;
 import com.woowacourse.thankoo.coupon.domain.MemberCoupon;
 import com.woowacourse.thankoo.coupon.exception.InvalidCouponException;
-import com.woowacourse.thankoo.coupon.infrastructure.MeetingProvider;
-import com.woowacourse.thankoo.coupon.infrastructure.ReservationProvider;
+import com.woowacourse.thankoo.coupon.domain.MeetingProvider;
+import com.woowacourse.thankoo.coupon.domain.ReservationProvider;
 import com.woowacourse.thankoo.coupon.presentation.dto.CouponDetailResponse;
 import com.woowacourse.thankoo.coupon.presentation.dto.CouponResponse;
 import com.woowacourse.thankoo.member.exception.InvalidMemberException;
